requirements.txt
# ==============================================================================
# Smart Customer Assistant MVP - Requirements
# ==============================================================================
# Core Dependencies for Multi-Agent AI System using AWS Strands
# ==============================================================================

# ------------------------------------------------------------------------------
# Strands Agents SDK (AWS Multi-Agent Framework)
# ------------------------------------------------------------------------------
strands-agents>=1.19.0
strands-agents-tools>=0.1.0

# ------------------------------------------------------------------------------
# AWS SDK for Bedrock Integration
# ------------------------------------------------------------------------------
boto3>=1.35.0
botocore>=1.35.0

# ------------------------------------------------------------------------------
# PyTorch for Recommendation Model (CPU version for lightweight demo)
# ------------------------------------------------------------------------------
torch>=2.0.0
numpy>=1.24.0

# ------------------------------------------------------------------------------
# Streamlit for MVP Frontend
# ------------------------------------------------------------------------------
streamlit>=1.29.0

# ------------------------------------------------------------------------------
# LangChain + LangGraph (Agentic AI Framework)
# ------------------------------------------------------------------------------
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-aws>=0.2.0
langgraph>=0.2.0

# ------------------------------------------------------------------------------
# Utilities
# ------------------------------------------------------------------------------
python-dotenv>=1.0.0
pydantic>=2.0.0

# ------------------------------------------------------------------------------
# Testing
# ------------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
